feat(setup): renamed files
