+++ src/navigation/stacks/public-stack.navigator.tsx
