+++ src/screens/Main/Main.screen.tsx
@@ -4,7 +4,7 @@ import { useState } from '@hookstate/core';
 import { AuthStore } from '@src/store/auth.store';
 import { BaseButton } from '@components/BaseButton';
 
-import { Wrapper, HeaderText } from './Main.styles';
+import { Wrapper, HeaderText } from './main.styles';
 
 const MainScreen = () => {
   const { user } = useState(AuthStore.state).get();
