+++ src/screens/SignIn/index.ts
@@ -1,1 +1,1 @@
-export { SignInScreen } from './SignIn.screen';
+export { SignInScreen } from './sign-in.screen';
